btrfs/011: fill the scratch device a little more
btrfs/011 is very flakey for me in my nightly automated tests, generally
failing because the device replace cancel races with the device replace
and tries to cancel after the replace has already finished.  To
accommodate faster machines, let the filler thread run a little longer,
so that there's enough work for device replace to do in order to be
successfully cancelled.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>