fstests: btrfs/011 mkfs the scratch dev before exiting
When testing encryption I started getting failures because the scratch
dev didn't have a valid fs at the end of the test.  This is because for
encryption we have to disable raid5/6, which changes how the test is
run.

Normally with raid6 we end up cancelling the device replace, and thus
$SCRATCH_DEV has a valid file system on it.  However with raid5/6
disabled we end with a normal DUP profile, and the replace doesn't end
up cancelled, so $SCRATCH_DEV is wiped.  Then when the test finishes we
do the normal fsck and see that there's no fs on the $SCRATCH_DEV and
error.

This test does all the fsck'ing during the workout period, so we don't
need the final scratch check, simply re-make the $SCRATCH_DEV at the end
as it could have been replaced during the test.

Signed-off-by: Josef Bacik <josef@toxicpanda.com>