fstests: guess proper MIN_FSSIZE when it's not set
Setting MIN_FSSIZE is necssary to skip a test, which create a FS with
a too small size. It is especially necesarry for the zoned case. Without
that, there will be many failed tests.

Guess a proper MIN_FSSIZE if not set, and skip such tests.

Currently, it assumes that a FS will be formated with "-d single -m dup"
(default) option. That can still conver all the RAID/DUP cases except
"-d dup -m dup".

Signed-off-by: Naohiro Aota <naohiro.aota@wdc.com>